Ändert die Anzahl der Schritte von short auf int * Java kann mit Datentypen nicht richtig umgehen und schon (short)0 ist umständlich * die API wird vereinfacht, die meisten Programmierer benutzen so oder so schon int und der Cast fällt weg * wenn die Programmierer trotzdem short benutzen, dann ist die Nutzung ohne weiteren Aufwand möglich (da short -> int sich casten lässt)

+++ org/KarlKuebelSchule/KugelmatikLib/Choreographies/SineWave.java
@@ -31,7 +31,7 @@ public class SineWave implements IChoreography {
     }
 
     @Override
-    public short getHeight(int x, int y, long millis, ChoreographyManager choreographyManager) {
+    public int getHeight(int x, int y, long millis, ChoreographyManager choreographyManager) {
         float v = x;
         if (direction == WaveDirection.Y)
             v = y;
@@ -42,7 +42,7 @@ public class SineWave implements IChoreography {
         sinWave += 1; // in den Bereich [0, 2] verschieben
         sinWave /= 2; // normalisieren
 
-        return (short)(sinWave * Config.MaxHeight);
+        return (int)(sinWave * Config.MaxHeight);
     }
 
 }
